requirements.txt
langchain-anthropic==0.1.16
langchain-core==0.2.28
langgraph==0.1.19
ipykernel==6.29.5
pandas==2.2.2
python-dotenv==1.0.1
openpyxl==3.1.5